+++ src-docs/src/views/datagrid/styling/datagrid_styling_example.js
@@ -1,4 +1,4 @@
-import React, { Fragment } from 'react';
+import React from 'react';
 
 import { GuideSectionTypes } from '../../../components';
 import { EuiDataGrid, EuiCode } from '../../../../../src/components';
@@ -7,6 +7,9 @@ import { EuiDataGridStyle } from '!!prop-loader!../../../../../src/components/da
 
 import DataGridStyling from './styling';
 const dataGridStylingSource = require('!!raw-loader!./styling_grid');
+
+import DataGridRowClasses from './row_classes';
+const dataGridRowClassesSource = require('!!raw-loader!./row_classes');
 
 import DataGridDisplayCallbacks from './display_callbacks';
 const dataGridDisplayCallbacksSource = require('!!raw-loader!./display_callbacks');
@@ -28,6 +31,23 @@ const ${gridSnippets.gridStyle};
 />
 `;
 
+const dataGridRowClassesSnippet = `
+<EuiDataGrid
+  aria-label="Data grid with gridStyle.rowClasses set"
+  columns={columns}
+  columnVisibility={{ visibleColumns, setVisibleColumns }}
+  rowCount={rowCount}
+  renderCellValue={renderCellValue}
+  gridStyle={{
+    rowClasses: {
+      0: 'someClass', // the first row will receive the 'someClass' class
+      1: 'anotherClass',
+      5: 'etc',
+    }
+  }}
+/>
+`;
+
 export const DataGridStylingExample = {
   title: 'Data grid style & display',
   sections: [
@@ -40,7 +60,7 @@ export const DataGridStylingExample = {
         },
       ],
       text: (
-        <Fragment>
+        <>
           <p>
             Styling can be passed down to the grid through the{' '}
             <EuiCode>gridStyle</EuiCode> prop. It accepts an object that allows
@@ -56,7 +76,7 @@ export const DataGridStylingExample = {
             parent font size or elements that use units relative to the parent
             container.
           </p>
-        </Fragment>
+        </>
       ),
       props: {
         EuiDataGrid,
@@ -64,6 +84,35 @@ export const DataGridStylingExample = {
       },
       snippet: gridStyleSnippet,
       demo: <DataGridStyling />,
+    },
+    {
+      title: 'Grid row classes',
+      source: [
+        {
+          type: GuideSectionTypes.JS,
+          code: dataGridRowClassesSource,
+        },
+      ],
+      text: (
+        <>
+          <p>
+            Specific rows can be highlighted or otherwise have custom styling
+            passed to them via the
+            <EuiCode>gridStyle.rowClasses</EuiCode> prop. It accepts an object
+            associating the row&apos;s index with a class name string.
+          </p>
+          <p>
+            The example below sets a custom striped class on the 3rd row and
+            dynamically updates the <EuiCode>rowClasses</EuiCode> map when rows
+            are selected.
+          </p>
+        </>
+      ),
+      props: {
+        EuiDataGridStyle,
+      },
+      snippet: dataGridRowClassesSnippet,
+      demo: <DataGridRowClasses />,
     },
     ...dataGridRowHeightOptionsExample.sections,
     {