+++ src/components/accessibility/screen_reader_only/index.ts
@@ -8,3 +8,4 @@
 
 export type { EuiScreenReaderOnlyProps } from './screen_reader_only';
 export { EuiScreenReaderOnly } from './screen_reader_only';
+export { euiScreenReaderOnlyStyles } from './screen_reader_only.styles';